PYTHONRASP-2: added rssi

+++ www/serverTracking.py
@@ -11,9 +11,11 @@ dbuser = 'game'
 dbpass = 'h95d3T7SXFta'
 
 # smoothing window size
-n = 7
+n = 3
 # target value, counting from the end. affects lag, -2 recommended
-n_2 = -4  # int(np.floor(n/2)) #
+n_2 = -2  # int(np.floor(n/2)) #
+triangulate = False
+debug = False
 
 locations_room = {}
 
@@ -149,40 +151,42 @@ def new_loop():
 			j = np.argmin(mean_values[:, n_2])
 			mean_value = mean_values[j, n_2]
 			location = mac_map[mac][j]
-			location_first = location
-			mean_value_first = mean_value
 
-			if location in locations_room["1"]['stations']:
-				room = "1"
-			else:
-				room = "2"
+			if triangulate:
+				location_first = location
+				mean_value_first = mean_value
 
-			locs = locations_room[room]['locations']
-			stat = locations_room[room]['stations']
-			indexs_dist = []
-			indexs_locs = []
-			for i, loc in enumerate(stat):
-				try:
-					index_dist = mac_map[mac].index(loc)
-					indexs_locs.append(i)
-					indexs_dist.append(index_dist)
-				except:
-					pass
+				if location in locations_room["1"]['stations']:
+					room = "1"
+				else:
+					room = "2"
 
-			dist_select = mean_values[indexs_dist, n_2]
-			locs_select = locs[indexs_locs]
-			# Trilateration
-			# initial_location = locs_select.mean(axis=0)
-			initial_location = locs[stat.index(location)]
-			x = optimization(initial_location, locs_select, dist_select)
+				locs = locations_room[room]['locations']
+				stat = locations_room[room]['stations']
+				indexs_dist = []
+				indexs_locs = []
+				for i, loc in enumerate(stat):
+					try:
+						index_dist = mac_map[mac].index(loc)
+						indexs_locs.append(i)
+						indexs_dist.append(index_dist)
+					except:
+						pass
 
-			new_dist = cdist(x.reshape((1, 2)), locs)[0, :]
+				dist_select = mean_values[indexs_dist, n_2]
+				locs_select = locs[indexs_locs]
+				# Trilateration
+				# initial_location = locs_select.mean(axis=0)
+				initial_location = locs[stat.index(location)]
+				x = optimization(initial_location, locs_select, dist_select)
 
-			j = np.argmin(new_dist)
-			location = stat[j]
-			mean_value = new_dist[j]
+				new_dist = cdist(x.reshape((1, 2)), locs)[0, :]
+
+				j = np.argmin(new_dist)
+				location = stat[j]
+				mean_value = new_dist[j]
+
 			mean_value_str = "{:.2f}".format(mean_value)
-
 			# print(x)
 			# print(location, mean_value)
 			# print(location_first, location, mean_value-mean_value_first)
@@ -205,7 +209,8 @@ def new_loop():
 		except:
 			pass
 	delta_time = datetime.now() - timedelta(minutes=1)
-	sql_request_remove = "DELETE FROM trackers_value WHERE timestamp < %s;"
+	if not debug:
+		sql_request_remove = "DELETE FROM trackers_value WHERE timestamp < %s;"
 	c.execute(sql_request_remove, delta_time)
 	connection.close()
 
@@ -237,7 +242,9 @@ def optimization(initial_location, locations, distances):
 
 
 if __name__ == "__main__":
-	run_once()
+
+	if not debug:
+		run_once()
 	init_locations()
 	while True:
 		# loop()
